fix: API specs

+++ repository/cartRepository.js
@@ -1,9 +1,5 @@
 const { sequelize } = require('../models/index');
-const BaseRepository = require('./BaseRepository');
 const { User, Product, Cart } = require('../models');
-
-const userBase = new BaseRepository(User);
-const cartBase = new BaseRepository(Cart);
 
 const queryInterface = sequelize.getQueryInterface();
 
@@ -17,40 +13,55 @@ const options = {
 };
 
 const findAllCarts = (async () => {
-  const allCarts = await userBase.all(options);
+  const carts = await User.findAll(options);
+  const allCarts = [];
+  for (const cart of carts) {
+    if (cart.products.length) {
+      allCarts.push(cart);
+    }
+  }
   return allCarts;
 });
 
 const findCartByUserId = (async (user_id) => {
   options.where = { id: user_id };
-  const cart = await userBase.one(options);
+  const cart = await User.findOne(options);
+  if (!cart || !cart.products.length) return null;
   return cart;
 });
 
 const createCart = (async (cart) => {
   const { id, products } = cart;
+  const newProducts = [];
   for (const product of products) {
+    const { product_id } = product;
+    const productAdded = await Cart.findOne({
+      where: { user_id: id, product_id },
+    });
+    if (productAdded) return null;
     product.user_id = id;
+    if (product.quantity) newProducts.push(product);
   }
-  const newCart = await queryInterface.bulkInsert('carts', products);
+  const newCart = await queryInterface.bulkInsert('carts', newProducts);
   return newCart;
 });
 
-const updateCart = (async (cart, value) => {
-  const updatedCart = await cartBase.update(cart, value);
-  return updatedCart;
-});
-
 const removeCartById = (async (user_id) => {
-  await Cart.destroy({
+  const exists = await Cart.findOne({
     where: { user_id },
   });
+  if (exists) {
+    await Cart.destroy({
+      where: { user_id },
+    });
+  } else {
+    return null;
+  }
 });
 
 module.exports = {
   findCartByUserId,
   findAllCarts,
   createCart,
-  updateCart,
   removeCartById,
 };
